fixed 404 if front end tried to delete non existing type

+++ app/Http/Controllers/SeatTypeController.php
@@ -99,7 +99,9 @@ class SeatTypeController extends Controller
      */
     public function destroy(SeatType $seatType)
     {
-        $seatType->delete();
+        if ($seatType) {
+            $seatType->delete();
+        }
 
         return back()->with('success', 'You deleted the seat type successfully');
     }
